add warning if sp model folder was deleted

+++ src/transformers/models/mbart50/tokenization_mbart50_fast.py
@@ -270,7 +270,9 @@ class MBart50TokenizerFast(PreTrainedTokenizerFast):
             save_directory, (filename_prefix + "-" if filename_prefix else "") + VOCAB_FILES_NAMES["vocab_file"]
         )
 
-        if os.path.abspath(self.vocab_file) != os.path.abspath(out_vocab_file):
+        if os.path.abspath(self.vocab_file) != os.path.abspath(out_vocab_file) and os.path.isfile(self.vocab_file):
             copyfile(self.vocab_file, out_vocab_file)
+        elif not os.path.isfile(self.vocab_file):
+            logger.warn(f"The original {self.vocab_file} was deleted and will not be saved with the tokenizer.")
 
         return (out_vocab_file,)
